Separate vesc_hi node and vesc_hi libraries
