+++ vesc_hi/CMakeLists.txt
@@ -13,6 +13,7 @@ find_package(catkin REQUIRED COMPONENTS
 
 catkin_package(
   INCLUDE_DIRS include
+  LIBRARIES vesc_hi
   CATKIN_DEPENDS roscpp serial hardware_interface controller_manager vesc_driver
 )
 
@@ -22,9 +23,19 @@ include_directories(
   ${catkin_INCLUDE_DIRS}
 )
 
-add_executable(vesc_hi
+add_library(vesc_hi
   src/vesc_hi.cpp
   src/vesc_servo_controller.cpp
 )
 add_dependencies(vesc_hi ${catkin_EXPORTED_TARGETS})
 target_link_libraries(vesc_hi ${catkin_LIBRARIES})
+
+add_executable(vesc_hi_node
+  src/vesc_hi_node.cpp
+)
+add_dependencies(vesc_hi_node ${catkin_EXPORTED_TARGETS})
+target_link_libraries(vesc_hi_node ${catkin_LIBRARIES} vesc_hi)
+
+install(DIRECTORY include/
+  DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
+)
